Move some helpers to a separate file
